Verified that casts get found and properly demangled
